separate contract wrapper
