+++ eth2near/contract_wrapper/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "contract_wrapper"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+borsh = "0.9.3"
+futures = "0.3.21"
+async-std = "1.12.0"
+near-sdk = "4.0.0"
+near-jsonrpc-client = "0.4.0-beta.0"
+near-crypto = "0.14.0"
+near-primitives = "0.14.0"
+near-chain-configs = "0.14.0"
+near-jsonrpc-primitives = "0.14.0"
+tokio = { version = "1.1", features = ["rt", "macros"] }
+reqwest = { version = "0.11", features = ["blocking"] }
+serde_json = "1.0.74"
+serde = { version = "1.0", features = ["derive"] }
