+++ eth2near/contract_wrapper/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod contract_wrapper_trait;
+pub mod near_contract_wrapper;
